validation for Product

+++ Shop.Application/Commands and Queries/Products/Commands/CreateProduct/CreateProductCommand.cs
@@ -13,6 +13,6 @@ namespace Shop.Application.Products.Commands.CreateProduct
         public decimal Price { get; set; }
         public string Description { get; set; }
         public int? OnStorageCount { get; set; }
-        public int? Rating { get; set; }
+        public double? Rating { get; set; }
     }
 }
