+++ Shop.Application/Commands and Queries/Products/Commands/DeleteProduct/DeleteProductCommand.cs
@@ -10,6 +10,6 @@ namespace Shop.Application.Products.Commands.DeleteProduct
         public decimal Price { get; set; }
         public string Description { get; set; }
         public int? OnStorageCount { get; set; }
-        public int? Rating { get; set; }
+        public double? Rating { get; set; }
     }
 }
